move 'utilities' packages into 'pkg'
