+++ internal/generic/resource_selector.go
@@ -9,8 +9,8 @@ import (
 	configv1alpha1 "github.com/common-fate/sdk/gen/commonfate/control/config/v1alpha1"
 	entityv1alpha1 "github.com/common-fate/sdk/gen/commonfate/entity/v1alpha1"
 	"github.com/common-fate/sdk/service/control/configsvc"
-	"github.com/common-fate/terraform-provider-commonfate/internal/utilities/diags"
-	"github.com/common-fate/terraform-provider-commonfate/internal/utilities/eid"
+	"github.com/common-fate/terraform-provider-commonfate/pkg/diags"
+	"github.com/common-fate/terraform-provider-commonfate/pkg/eid"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
